TF 1.6.0 Errors fixed
