+++ tensorlayer/decorators/deprecated.py
@@ -5,9 +5,9 @@ import inspect
 import sys
 import functools
 
-from tensorlayer.decorators.utils.decorator_utils import get_qualified_name
-from tensorlayer.decorators.utils.decorator_utils import _normalize_docstring
-from tensorlayer.decorators.utils.deprecation_utils import _validate_deprecation_args
+from tensorlayer.decorators.utils import add_deprecation_notice_to_docstring
+from tensorlayer.decorators.utils import get_qualified_name
+from tensorlayer.decorators.utils import validate_deprecation_args
 
 import wrapt
 
@@ -20,49 +20,6 @@ _PRINT_DEPRECATION_WARNINGS = True
 _PRINTED_WARNING = {}
 
 
-def add_notice_to_docstring(doc, no_doc_str, notice):
-    """Adds a deprecation notice to a docstring."""
-    if not doc:
-        lines = [no_doc_str]
-
-    else:
-        lines = _normalize_docstring(doc).splitlines()
-
-    notice = [''] + notice
-
-    if len(lines) > 1:
-        # Make sure that we keep our distance from the main body
-        if lines[1].strip():
-            notice.append('')
-
-        lines[1:1] = notice
-    else:
-        lines += notice
-
-    return '\n'.join(lines)
-
-
-def _add_deprecated_function_notice_to_docstring(doc, date, instructions):
-    """Adds a deprecation notice to a docstring for deprecated functions."""
-
-    if instructions:
-        deprecation_message = """
-            .. warning::
-                **THIS FUNCTION IS DEPRECATED:** It will be removed after %s.
-                *Instructions for updating:* %s.
-        """ % (('in a future version' if date is None else ('after %s' % date)), instructions)
-
-    else:
-        deprecation_message = """
-            .. warning::
-                **THIS FUNCTION IS DEPRECATED:** It will be removed after %s.
-        """ % (('in a future version' if date is None else ('after %s' % date)))
-
-    main_text = [deprecation_message]
-
-    return add_notice_to_docstring(doc, 'DEPRECATED FUNCTION', main_text)
-
-
 def deprecated(wrapped=None, date='', instructions='', warn_once=True):
 
     if wrapped is None:
@@ -71,7 +28,7 @@ def deprecated(wrapped=None, date='', instructions='', warn_once=True):
     @wrapt.decorator
     def wrapper(wrapped, instance=None, args=None, kwargs=None):
 
-        _validate_deprecation_args(date, instructions)
+        validate_deprecation_args(date, instructions)
 
         if _PRINT_DEPRECATION_WARNINGS:
 
@@ -98,7 +55,7 @@ def deprecated(wrapped=None, date='', instructions='', warn_once=True):
 
     if sys.version_info > (3, 0):  # docstring can only be edited with Python 3
         wrapt.FunctionWrapper.__setattr__(
-            decorated, "__doc__", _add_deprecated_function_notice_to_docstring(wrapped.__doc__, date, instructions)
+            decorated, "__doc__", add_deprecation_notice_to_docstring(wrapped.__doc__, date, instructions)
         )
 
     return decorated
