--- tensorlayer/decorators/utils/__init__.py
@@ -1,14 +1,0 @@
-#! /usr/bin/python
-# -*- coding: utf-8 -*-
-"""
-NOTE: DO NOT REMOVE THESE FILES. They are copied from Tensorflow repository and are necessary to build the library without installing TF.
-
-Source: https://github.com/tensorflow/tensorflow/tree/master/tensorflow/python/util
-
-They replace the following imports:
->>> from tensorflow.python.util import decorator_utils
->>> from tensorflow.python.util.deprecation import _validate_deprecation_args
-"""
-
-from tensorlayer.decorators.utils.decorator_utils import *
-from tensorlayer.decorators.utils.deprecation_utils import *
